Update FINAL_SRUC MOU R Script (NotePad).r
Implements code that will subset the FeeStatus data frame based on part time status, and export that to part of the outputs folder so that it can be used as an input template for next year. This is a way to ensure that part-time students are still included in the FeeStatus data in the 2nd and 3rd years of their programmes.